+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+* Implement support for SVG images.
+    - Add `zng-ext-svg`.
+    - Add non default `"svg"` feature on the `zng` crate.
 * Fix `view_process_extension!` not running in same-process mode.
 * **Breaking** Add `as_any` casting method for `RendererExtension` and `WindowExtension`.
     - Note that this is only a breaking change for direct dependents of `zng-view` extensions API.